fix: fix type errors
