+++ packages/core/src/common/interceptor.ts
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
+import type { Nullable } from '../shared/types';
 import { remove } from './array';
-import type { Nullable } from './type-util';
 
 export type InterceptorHandler<M = unknown, C = unknown> = (
     value: Nullable<M>,
