+++ packages/core/src/services/permission/type.ts
@@ -17,9 +17,9 @@
 import type { BehaviorSubject, Observable } from 'rxjs';
 import type { UnitAction, UnitObject } from '@univerjs/protocol';
 import { createIdentifier } from '../../common/di';
-import type { Nullable } from '../../common/type-utils';
 import { LifecycleStages, runOnLifecycle } from '../lifecycle/lifecycle';
-import type { IRange } from '../../types/interfaces';
+import type { IRange } from '../../sheets/typedef';
+import type { Nullable } from '../../shared/types';
 
 export enum PermissionStatus {
     INIT = 'init',
