+++ packages/core/src/services/theme/theme.service.ts
@@ -17,8 +17,8 @@
 import type { Observable } from 'rxjs';
 import { BehaviorSubject } from 'rxjs';
 
-import type { Nullable } from '../../common/type-util';
 import { Disposable, toDisposable } from '../../shared/lifecycle';
+import type { Nullable } from '../../shared/types';
 
 export interface IStyleSheet {
     [key: string]: string;
