+++ packages/core/src/types/interfaces/i-cell-custom-render.ts
@@ -19,9 +19,8 @@
 import type { Nullable } from '../../shared';
 import type { Worksheet } from '../../sheets/worksheet';
 import type { Workbook } from '../../sheets/workbook';
-import type { ISelectionCellWithMergeInfo } from './i-selection-data';
+import type { ICellDataForSheetInterceptor, ISelectionCellWithMergeInfo } from '../../sheets/typedef';
 import type { IStyleData } from './i-style-data';
-import type { ICellDataForSheetInterceptor } from './i-cell-data';
 
 export interface ICellRenderContext {
     data: ICellDataForSheetInterceptor;
