+++ packages/core/src/types/interfaces/i-document-data.ts
@@ -14,17 +14,16 @@
  * limitations under the License.
  */
 
-import type { Nullable } from '../../common/type-util';
 import type { IAbsoluteTransform, ISize } from '../../shared/shape';
+import type { Nullable } from '../../shared/types';
 import type { BooleanNumber, CellValueType, HorizontalAlign, LocaleType, TextDirection, VerticalAlign, WrapStrategy } from '../enum';
-import type { IExtraModelData } from './i-extra-model-data';
 import type { IColorStyle, IStyleBase } from './i-style-data';
 
 // Attention: all dimensional units, unless otherwise stated, refer to pt，1 pt = 1 / 72 in
 /**
  * Properties of document
  */
-export interface IDocumentData extends IReferenceSource, IExtraModelData {
+export interface IDocumentData extends IReferenceSource {
     /** unit ID */
     id: string;
     /** Revision of this document. Would be used in collaborated editing. Starts with zero. */
