+++ packages/core/src/types/interfaces/i-slide-data.ts
@@ -16,6 +16,7 @@
 
 import type { ISize } from '../../shared/shape';
 import type { IKeyType, Nullable } from '../../shared/types';
+import type { IWorksheetData } from '../../sheets/typedef';
 import type { LocaleType, ThemeColorType } from '../enum';
 import type { ShapeType } from '../enum/prst-geom-type';
 import type { ICustomBlock, IDocumentData, ILists, ITransformState } from './i-document-data';
@@ -23,7 +24,6 @@ import type { IImageProperties } from './i-image-properties';
 import type { IPlaceholder } from './i-placeholder';
 import type { IShapeProperties } from './i-shape-properties';
 import type { IColorStyle, IStyleBase, IStyleData } from './i-style-data';
-import type { IWorksheetData } from './i-worksheet-data';
 
 export interface ISlideData extends IReferenceSource {
     id: string; // unit id
@@ -116,11 +116,15 @@ export interface IPageElement {
     shape?: IShape;
     image?: IImage;
     richText?: IRichTextProps;
+
+    /** @deprecated */
     spreadsheet?: {
         worksheet: IWorksheetData;
         styles: IKeyType<Nullable<IStyleData>>;
     };
+    /** @deprecated */
     document?: IDocumentData;
+    /** @deprecated */
     slide?: ISlideData;
     // video: IVideo;
     // line: ILine;
